Autonomous almost done
