+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/AutoTest.java
@@ -1,0 +1,64 @@
+package org.firstinspires.ftc.teamcode;
+
+import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+import com.qualcomm.robotcore.hardware.DcMotor;
+import com.qualcomm.robotcore.hardware.NormalizedColorSensor;
+import com.qualcomm.robotcore.hardware.Servo;
+
+/**
+ * Created by pramo on 12/14/2017.
+ */
+
+@Autonomous
+public class AutoTest extends LinearOpMode{
+
+    private DcMotor motor0;
+    private DcMotor motor1;
+    private DcMotor motor2;
+    private DcMotor motor3;
+    private Servo servo0;
+    private Servo servo1;
+    private Servo servo2;
+    private NormalizedColorSensor colorSensor;
+    private TankDriveTrain driveTrain;
+    private Grabber grabber;
+    private VerticalLiftMotor liftMotor;
+    private Autonomous_Code Auto;
+
+    @Override
+    public void runOpMode() {
+
+        initialization();
+        Auto = new Autonomous_Code(motor0, motor1, motor2, motor3, servo0, servo1, servo2, colorSensor, driveTrain, grabber, liftMotor);
+
+        //Wait for the game to start (driver presses PLAY)
+        waitForStart();
+
+        //run until the end of the match (driver presses STOP)
+        while (opModeIsActive()) {
+
+            Auto.auto("back", "red");
+            stop();
+
+        }
+    }
+
+    public void initialization() {
+        //HARDWARE MAPS
+        motor0 = hardwareMap.get(DcMotor.class, "motor0");
+        motor1 = hardwareMap.get(DcMotor.class, "motor1");
+        motor2 = hardwareMap.get(DcMotor.class, "motor2");
+        // motor3 = hardwareMap.get(DcMotor.class, "motor3");
+        servo0 = hardwareMap.get(Servo.class, "servo0");
+        servo1 = hardwareMap.get(Servo.class, "servo1");
+        servo2 = hardwareMap.get(Servo.class, "servo2");
+        colorSensor = hardwareMap.get(NormalizedColorSensor.class, "colorSensor");
+
+        //INITIALIZATION
+        driveTrain = new TankDriveTrain(motor0, motor1);
+        grabber = new Grabber(servo1, servo2);
+        liftMotor = new VerticalLiftMotor(motor2);
+    }
+
+}
